refactor: dto 이름 통일 및 패키지 세분화
