+++ src/main/java/com/profect/tickle/domain/reservation/dto/response/reservation/PaymentInfo.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.reservation.dto.response;
+package com.profect.tickle.domain.reservation.dto.response.reservation;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
