+++ src/main/java/com/profect/tickle/domain/reservation/dto/response/reservation/PerformanceInfo.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.reservation.dto.response;
+package com.profect.tickle.domain.reservation.dto.response.reservation;
 
 import java.time.Instant;
 import lombok.AllArgsConstructor;
