+++ src/main/java/com/profect/tickle/domain/reservation/dto/response/reservation/ReservationDetailResponseDto.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.reservation.dto.response;
+package com.profect.tickle.domain.reservation.dto.response.reservation;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import java.util.List;
 @Getter
 @Builder
 @AllArgsConstructor
-public class ReservationDetailResponse {
+public class ReservationDetailResponseDto {
     private Long reservationId;
     private String reservationNumber;
