+++ src/main/java/com/profect/tickle/domain/reservation/dto/response/reservation/ReservationInfoResponseDto.java
@@ -1,6 +1,7 @@
-package com.profect.tickle.domain.reservation.dto.response;
+package com.profect.tickle.domain.reservation.dto.response.reservation;
 
 import com.profect.tickle.domain.event.dto.response.CouponResponseDto;
+import com.profect.tickle.domain.reservation.dto.response.preemption.PreemptedSeatInfo;
 import java.time.Instant;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class ReservationInfoResponse {
+public class ReservationInfoResponseDto {
     private List<PreemptedSeatInfo> seats;         // 선점된 좌석 정보
     private Integer totalAmount;                   // 총 결제 금액
     private Integer currentPoints;                 // 보유 포인트
