+++ src/main/java/com/profect/tickle/domain/reservation/dto/response/reservation/ReservedSeatDetail.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.reservation.dto.response;
+package com.profect.tickle.domain.reservation.dto.response.reservation;
 
 import com.profect.tickle.domain.reservation.entity.SeatGrade;
 import lombok.AllArgsConstructor;
